Fixing data types in tests.

+++ Test/Core/CoreSystemTests.hpp
@@ -41,9 +41,9 @@ namespace Curse
             EXPECT_EQ(Seconds<float>(2.0f).AsSeconds<float>(), 2.0f);
             EXPECT_EQ(Seconds<float>(-1.0f).AsSeconds<float>(), -1.0f);
             EXPECT_EQ(Seconds<float>(-2.0f).AsSeconds<float>(), -2.0f);
-            EXPECT_EQ(Seconds<int32_t>(12345).AsSeconds<int32_t>(), 12345);
-            EXPECT_EQ(Seconds<int32_t>(-12345).AsSeconds<int32_t>(), -12345);
-            EXPECT_EQ(Seconds<float>(1.5f).AsSeconds<int32_t>(), 1);
+            EXPECT_EQ(Seconds<int32_t>(12345).AsSeconds<int32_t>(), int32_t(12345));
+            EXPECT_EQ(Seconds<int32_t>(-12345).AsSeconds<int32_t>(), int32_t(-12345));
+            EXPECT_EQ(Seconds<float>(1.5f).AsSeconds<int32_t>(), int32_t(1));
   
             EXPECT_EQ(Seconds<int64_t>(2).AsSeconds<int64_t>(), 2LL);
             EXPECT_EQ(Seconds<int64_t>(2).AsMilliseconds<int64_t>(), 2000LL);
@@ -55,62 +55,62 @@ namespace Curse
             EXPECT_EQ(Milliseconds<float>(2.0f).AsSeconds<float>(), 0.002f);
             EXPECT_EQ(Milliseconds<float>(-1.0f).AsSeconds<float>(), -0.001f);
             EXPECT_EQ(Milliseconds<float>(-2.0f).AsSeconds<float>(), -0.002f);
-            EXPECT_EQ(Milliseconds<int32_t>(12345).AsSeconds<int32_t>(), 12);
-            EXPECT_EQ(Milliseconds<int32_t>(-12345).AsSeconds<int32_t>(), -12);
-            EXPECT_EQ(Milliseconds<float>(1000.5f).AsSeconds<int32_t>(), 1);
+            EXPECT_EQ(Milliseconds<int32_t>(12345).AsSeconds<int32_t>(), int32_t(12));
+            EXPECT_EQ(Milliseconds<int32_t>(-12345).AsSeconds<int32_t>(), int32_t(-12));
+            EXPECT_EQ(Milliseconds<float>(1000.5f).AsSeconds<int32_t>(), int32_t(1));
         }
         {
             EXPECT_EQ(Microseconds<float>(1.0f).AsMilliseconds<float>(), 0.001f);
             EXPECT_EQ(Microseconds<float>(2.0f).AsMilliseconds<float>(), 0.002f);
             EXPECT_EQ(Microseconds<float>(-1.0f).AsMilliseconds<float>(), -0.001f);
             EXPECT_EQ(Microseconds<float>(-2.0f).AsMilliseconds<float>(), -0.002f);
-            EXPECT_EQ(Microseconds<int32_t>(12345).AsMilliseconds<int32_t>(), 12);
-            EXPECT_EQ(Microseconds<int32_t>(-12345).AsMilliseconds<int32_t>(), -12);
-            EXPECT_EQ(Microseconds<float>(1000.5f).AsMilliseconds<int32_t>(), 1);
+            EXPECT_EQ(Microseconds<int32_t>(12345).AsMilliseconds<int32_t>(), int32_t(12));
+            EXPECT_EQ(Microseconds<int32_t>(-12345).AsMilliseconds<int32_t>(), int32_t(-12));
+            EXPECT_EQ(Microseconds<float>(1000.5f).AsMilliseconds<int32_t>(), int32_t(1));
         }
         {
             EXPECT_EQ(Nanoseconds<float>(1.0f).AsMicroseconds<float>(), 0.001f);
             EXPECT_EQ(Nanoseconds<float>(2.0f).AsMicroseconds<float>(), 0.002f);
             EXPECT_EQ(Nanoseconds<float>(-1.0f).AsMicroseconds<float>(), -0.001f);
             EXPECT_EQ(Nanoseconds<float>(-2.0f).AsMicroseconds<float>(), -0.002f);
-            EXPECT_EQ(Nanoseconds<int32_t>(12345).AsMicroseconds<int32_t>(), 12);
-            EXPECT_EQ(Nanoseconds<int32_t>(-12345).AsMicroseconds<int32_t>(), -12);
-            EXPECT_EQ(Nanoseconds<float>(1000.5f).AsMicroseconds<int32_t>(), 1);
+            EXPECT_EQ(Nanoseconds<int32_t>(12345).AsMicroseconds<int32_t>(), int32_t(12));
+            EXPECT_EQ(Nanoseconds<int32_t>(-12345).AsMicroseconds<int32_t>(), int32_t(-12));
+            EXPECT_EQ(Nanoseconds<float>(1000.5f).AsMicroseconds<int32_t>(), int32_t(1));
         }
     }
     TEST(System, Time_Operators)
     {
         {
             Time time = Seconds<int32_t>(3000) - Seconds<int32_t>(1000);
-            EXPECT_EQ(time.AsSeconds<int32_t>(), 2000);
+            EXPECT_EQ(time.AsSeconds<int32_t>(), int32_t(2000));
         }
         {
             Time time = Seconds<int32_t>(5000) - Seconds<int32_t>(6000);
-            EXPECT_EQ(time.AsSeconds<int32_t>(), -1000);
+            EXPECT_EQ(time.AsSeconds<int32_t>(), int32_t(-1000));
         }
         {
             Time time = Seconds<int32_t>(5000) - Seconds<int32_t>(-6000);
-            EXPECT_EQ(time.AsSeconds<int32_t>(), 11000);
+            EXPECT_EQ(time.AsSeconds<int32_t>(), int32_t(11000));
         }
         {
             Time time = Seconds<int32_t>(3000) + Seconds<int32_t>(1000);
-            EXPECT_EQ(time.AsSeconds<int32_t>(), 4000);
+            EXPECT_EQ(time.AsSeconds<int32_t>(), int32_t(4000));
         }
         {
             Time time = Seconds<int32_t>(5000) + Seconds<int32_t>(6000);
-            EXPECT_EQ(time.AsSeconds<int32_t>(), 11000);
+            EXPECT_EQ(time.AsSeconds<int32_t>(), int32_t(11000));
         }
         {
             Time time = Seconds<int32_t>(5000) + Seconds<int32_t>(-6000);
-            EXPECT_EQ(time.AsSeconds<int32_t>(), -1000);
+            EXPECT_EQ(time.AsSeconds<int32_t>(), int32_t(-1000));
         }
         {
             Time time = Seconds<int32_t>(5) * 6.0f;
-            EXPECT_EQ(time.AsSeconds<int32_t>(), 30);
+            EXPECT_EQ(time.AsSeconds<int32_t>(), int32_t(30));
         }
         {
             Time time = Seconds<int32_t>(5) * -6.5f;
-            EXPECT_EQ(time.AsSeconds<int32_t>(), -32);
+            EXPECT_EQ(time.AsSeconds<int32_t>(), int32_t(-32));
         }
     }
 
@@ -118,11 +118,11 @@ namespace Curse
     {
         {
             Semaphore sem;
-            EXPECT_EQ(sem.GetWaitCount(), 0);
+            EXPECT_EQ(sem.GetWaitCount(), size_t(0));
             sem.NotifyOne();
-            EXPECT_EQ(sem.GetWaitCount(), 0);
+            EXPECT_EQ(sem.GetWaitCount(), size_t(0));
             sem.Wait();
-            EXPECT_EQ(sem.GetWaitCount(), 0);
+            EXPECT_EQ(sem.GetWaitCount(), size_t(0));
         }
     }
     TEST(System, Semaphore_NotifyOne)
@@ -134,14 +134,14 @@ namespace Curse
                 [&sem]()
                 {
                     std::this_thread::sleep_for(std::chrono::duration<double>(0.5f));
-                    EXPECT_EQ(sem.GetWaitCount(), 1);
+                    EXPECT_EQ(sem.GetWaitCount(), size_t(1));
                     sem.NotifyOne();
                 });
     
             sem.Wait();
             thread.join();
 
-            EXPECT_EQ(sem.GetWaitCount(), 0);
+            EXPECT_EQ(sem.GetWaitCount(), size_t(0));
         }
     }
     TEST(System, Semaphore_NotifyAll)
@@ -160,7 +160,7 @@ namespace Curse
             };
 
 
-            EXPECT_EQ(sem.GetWaitCount(), 0);
+            EXPECT_EQ(sem.GetWaitCount(), size_t(0));
             for (size_t i = 0; i < thread_count; i++)
             {
                 threads[i] = std::thread(func, std::ref(status[i]));
@@ -176,7 +176,7 @@ namespace Curse
                 EXPECT_TRUE(status[i]);
             }
 
-            EXPECT_EQ(sem.GetWaitCount(), 0);
+            EXPECT_EQ(sem.GetWaitCount(), size_t(0));
         }
     }
